+++ Apps/pisowise-fe/app/projects/[id]/components/cards/AIInsightsCard.tsx
